+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -169,7 +169,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-24T13:24:31.773Z",
+    "2026-06-24T14:42:39.367Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {