+++ packages/orq-rc/RELEASES.md
@@ -10239,3 +10239,13 @@ Based on:
 - [typescript v4.11.0-rc.45] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.45] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.45 - packages/orq-rc
+
+## 2026-06-24 14:45:11
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.785.0 (2.912.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.46] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.46] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.46 - packages/orq-rc